pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ra-rom-manager"
version = "0.1.0"
description = "Personal RetroAchievements ROM manager — hash verification, achievement tracking, and Excel reporting"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.9",
    "types-requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # do not perform function calls in default arguments
]

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]